requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
python-dotenv==1.0.0
itsdangerous==2.1.2
gunicorn==20.1.0